requirements.txt
fastapi
uvicorn
python-dotenv
openai
google-generativeai
tavily-python
beautifulsoup4
replicate
